requirements.txt
python-dotenv==1.0.0
telethon>=1.26.0
aiohttp>=3.10.10
pillow~=11.0.0